requirements.txt
langchain
langchain-community
langchain-openai
langchain-core
langchain-anthropic
langchain-google-genai
langchain-groq
langchain-ollama
langchainhub
black # for formatting
python-dotenv # for loading environment variables from .env file
langchain-pinecone
langchain-huggingface
sentence_transformers
faiss-cpu # for vector store
pypdf
azure-search-documents
azure-identity
streamlit
